erp-bot/requirements.txt
discord.py>=2.4.0
groq
python-dotenv